vello_common

Module clip

Managing clipping state.

Structs

ClipContext

A context for managing clip stacks.

ClipData 🔒

ClipFrame 🔒

A frame containing clipping-relevant state for the root layer or a filter layer.

ClipState

State for managing clip paths across multiple viewports.

FillRegion 🔒

Overlap 🔒

An overlap between two regions.

PathDataRef

Borrowed data of a stripped path.

RawClip 🔒

Raw data of a previously pushed clip path.

RowIterator 🔒

An iterator of strip and fill regions of a single strip row.

StripRegion 🔒

StripState 🔒

The data of the current strip we are building.

Enums

Advance 🔒

OverlapRelationship 🔒

The relationship between two regions.

Region 🔒

Functions

first_strip_at_or_after 🔒

flush_strip 🔒

intersect

Compute the sparse strips representation of a path that results from intersecting the two input paths. This can be used to implement clip paths.

intersect_impl 🔒

The implementation of the clipping algorithm using sparse strips. Conceptually, it is relatively simple: We iterate over each strip and fill region of the two paths in lock step and determine all overlaps between the two. For each overlap, we proceed depending on what kind of region we have in the first path and the second one.

should_create_new_strip 🔒

start_strip 🔒