vello_common

Module coarse

Generating and processing wide tiles.

Structs

Bbox 🔒

A bounding box

Clip 🔒

A clip region.

CmdAlphaFill

Fill a consecutive region of a wide tile with an alpha mask.

CmdClipAlphaFill

Same as strip, but composites top of clip stack to next on stack.

CmdClipFill

Same as fill, but copies top of clip stack to next on stack.

CmdFill

Fill a consecutive region of a wide tile.

Layer 🔒

Wide

A container for wide tiles.

WideTile

A wide tile.

Enums

Cmd

A drawing command.

Constants

MODE_CPU

MODE_CPU allows compile time optimizations to be applied to wide tile draw command generation specific to vello_cpu.

MODE_HYBRID

MODE_HYBRID allows compile time optimizations to be applied to wide tile draw command generation specific for vello_hybrid.

Traits

BlendModeExt 🔒