vello_common::coarse

Struct WideTile

pub struct WideTile<const MODE: u8 = MODE_CPU> {
    pub x: u16,
    pub y: u16,
    pub bg: PremulColor,
    pub cmds: Vec<Cmd>,
    pub n_zero_clip: usize,
    pub n_clip: usize,
    pub n_bufs: usize,
}

A wide tile.

Fields
x: u16

The x coordinate of the wide tile.

y: u16

The y coordinate of the wide tile.

bg: PremulColor

The background of the tile.

cmds: Vec<Cmd>

The draw commands of the tile.

n_zero_clip: usize

The number of zero-winding clips.

n_clip: usize

The number of non-zero-winding clips.

n_bufs: usize

The number of pushed buffers.

Implementations

impl WideTile

pub const WIDTH: u16 = 256u16

The width of a wide tile in pixels.

pub const MAX_WIDE_TILE_COORD: u16 = 255u16

The maximum coordinate of a wide tile.

impl WideTile<MODE_CPU>

pub fn new(width: u16, height: u16) -> Self

Create a new wide tile.

impl WideTile<MODE_HYBRID>

pub fn new(width: u16, height: u16) -> Self

Create a new wide tile.

impl<const MODE: u8> WideTile<MODE>

fn new_internal(x: u16, y: u16) -> Self

Create a new wide tile.

pub(crate) fn fill(&mut self, x: u16, width: u16, paint: Paint)

pub(crate) fn strip(&mut self, cmd_strip: CmdAlphaFill)

pub fn push_clip(&mut self)

Adds a new clip region to the current wide tile.

pub fn pop_clip(&mut self)

Removes the most recently added clip region from the current wide tile.

pub fn push_zero_clip(&mut self)

Adds a zero-winding clip region to the stack.

pub fn pop_zero_clip(&mut self)

Removes the most recently added zero-winding clip region.

pub fn is_zero_clip(&mut self) -> bool

Checks if the current clip region is a zero-winding clip.

pub fn clip_strip(&mut self, cmd_clip_strip: CmdClipAlphaFill)

Applies a clip strip operation with the given parameters.

pub fn clip_fill(&mut self, x: u32, width: u32)

Applies a clip fill operation at the specified position and width.

pub fn push_buf(&mut self)

Push a buffer.

pub fn pop_buf(&mut self)

Pop the most recent buffer.

pub fn opacity(&mut self, opacity: f32)

Apply an opacity to the whole buffer.

pub fn mask(&mut self, mask: Mask)

Apply a mask to the whole buffer.

pub fn blend(&mut self, blend_mode: BlendMode)

Blend the current buffer into the previous buffer in the stack.