vello_common::colr

Trait ColrRenderer

pub trait ColrRenderer {
    // Required methods
    fn push_clip_layer(&mut self, clip: &BezPath);
    fn push_blend_layer(&mut self, blend_mode: BlendMode);
    fn fill_solid(&mut self, color: AlphaColor<Srgb>);
    fn fill_gradient(&mut self, gradient: Gradient);
    fn set_paint_transform(&mut self, affine: Affine);
    fn pop_layer(&mut self);
}

A trait for clients capable of rendering COLR glyphs.

Required Methods

fn push_clip_layer(&mut self, clip: &BezPath)

Push a new clip layer.

fn push_blend_layer(&mut self, blend_mode: BlendMode)

Push a new blend layer.

fn fill_solid(&mut self, color: AlphaColor<Srgb>)

Fill the current area with the given solid color.

fn fill_gradient(&mut self, gradient: Gradient)

Fill the current area with the given gradient color.

fn set_paint_transform(&mut self, affine: Affine)

Set the transform for paints.

fn pop_layer(&mut self)

Pop the last clip/blend layer.