vello_common

Module encode

Paints for drawing shapes.

Modules

private 🔒

Structs

EncodedBlurredRoundedRectangle

An encoded blurred, rounded rectangle.

EncodedExternalTexture

An encoded external texture.

EncodedGradient

An encoded gradient.

EncodedImage

An encoded image.

FocalData

Focal data for a radial gradient.

GradientCacheKey

Cache key for gradient color ramps based on color-affecting properties.

GradientLut

A lookup table for sampled gradient values.

GradientRange

An encoded range between two color stops.

LinearKind

Computed properties of a linear gradient.

SweepKind

Computed properties of a sweep gradient.

Enums

EncodedKind

A kind of encoded gradient.

EncodedPaint

An encoded paint.

RadialKind

A radial gradient.

Constants

DEGENERATE_THRESHOLD 🔒

MAX_GRADIENT_LUT_SIZE

The maximum size of the gradient LUT.

NUDGE_VAL 🔒

Traits

EncodeExt

A trait for encoding paints.

GradientLutExt

A helper trait for converting gradient colors to Self.

Functions

determine_lut_size 🔒

encode_stops 🔒

Encode all stops into a sequence of ranges.

exp 🔒

ts_from_line_to_line 🔒

Calculates the transform necessary to map the line spanned by points src1, src2 to the line spanned by dst1, dst2.

unit_to_line 🔒

Calculate the transform necessary to map the unit vector to the line spanned by the points p1 and p2.

validate 🔒

Returns a fallback paint in case the gradient is invalid.

x_y_advances 🔒