vello_common

Module filter

Common filter helper functions.

Unlike the filters defines in crate::filter_effects, the filters in this module represent a special representation of each filter to be used as the basis for rendering in vello_hybrid and vello_cpu.

Modules

drop_shadow

The drop shadow filter.

flood

The flood filter.

gaussian_blur

The gaussian blur filter.

offset

The offset filter.

Structs

FilterData

Precomputed data for a filter layer.

FilterLayerPlacement

Metadata about a filter layer and how it should be composited back into the parent layer.

Enums

PreparedFilter

A filter that has been prepared for rendering.

Functions

transform_offset_params 🔒

Transform an offset’s dx/dy using the affine transformation’s linear part.