vello_common

Module flatten_simd

This is a temporary module that contains a SIMD version of flattening of cubic curves, as well as some code that was copied from kurbo, which is needed to reimplement the full flatten method.

Structs

FlattenCtx

The context needed for flattening curves.

FlattenParams ğŸ”’

Point32 ğŸ”’

Constants

MAX_QUADS ğŸ”’

This limit was chosen based on the pre-existing GitHub gist. This limit should not be hit in normal operation, but might be hit for very large transforms.

TO_QUAD_TOL ğŸ”’

Traits

Callback ğŸ”’

FlattenParamsExt ğŸ”’

Functions

approx_parabola_integral ğŸ”’

An approximation to $\int (1 + 4x^2) ^ -0.25 dx$

approx_parabola_integral_simd ğŸ”’

approx_parabola_integral_simd_x4 ğŸ”’

approx_parabola_inv_integral ğŸ”’

An approximation to the inverse parabola integral.

approx_parabola_inv_integral_simd ğŸ”’

estimate ğŸ”’

estimate_num_quads ğŸ”’

estimate_subdiv_simd ğŸ”’

eval_cubics_simd ğŸ”’

eval_simd ğŸ”’

flatten ğŸ”’

See the docs for the kurbo implementation of flattening: https://docs.rs/kurbo/latest/kurbo/fn.flatten.html

flatten_cubic_simd ğŸ”’

is_finite_simd ğŸ”’

output_lines_simd ğŸ”’

pt_splat_simd ğŸ”’