vello_common

Module glyph

Processing and drawing glyphs.

Structs

BitmapGlyph

A glyph defined by a bitmap.

ColorGlyph

A glyph defined by a COLR glyph description.

Glyph

Positioned glyph.

GlyphCaches

Caches used for glyph rendering.

GlyphRun 🔒

A sequence of glyphs with shared rendering properties.

GlyphRunBuilder

A builder for configuring and drawing glyphs.

HintCache 🔒

LRU cache for hinting instances.

HintEntry 🔒

HintKey 🔒

OutlineCache 🔒

Caches glyph outlines for reuse. Heavily inspired by vello_encoding::glyph_cache.

OutlineCacheSession 🔒

OutlineEntry 🔒

OutlineGlyph

A glyph defined by a path (its outline) and a local transform.

OutlineKey 🔒

OutlinePath 🔒

PreparedGlyph

A simplified representation of a glyph, prepared for easy rendering.

PreparedGlyphRun 🔒

VarLookupKey 🔒

Lookup key for variable font caches.

Enums

GlyphType

A type of glyph.

Style

Rendering style for glyphs.

Constants

HINTING_OPTIONS 🔒

MAX_CACHED_HINT_INSTANCES 🔒

We keep this small to enable a simple LRU cache with a linear search. Regenerating hinting data is low to medium cost so it’s fine to redo it occasionally.

Traits

GlyphRenderer

Trait for types that can render glyphs.

Functions

find_hint_entry 🔒

prepare_bitmap_glyph 🔒

prepare_colr_glyph 🔒

prepare_glyph_run 🔒

Prepare a glyph run for rendering.

prepare_outline_glyph 🔒

Type Aliases

NormalizedCoord

A normalized variation coordinate (for variable fonts) in 2.14 fixed point format.

VarKey 🔒

Key for variable font caches.