vello_common

Module paint

Types for paints.

Structs

ImageId

Opaque image handle

IndexedPaint

A paint that needs to be resolved via its index.

NoOpImageResolver

A no-op image resolver that always returns None.

PremulColor

A premultiplied color.

Tint

A tint applied to image paints.

Enums

ImageSource

Bitmap source used by Image.

Paint

A paint used internally by a rendering frontend to store how a draw should be painted. There are only two types of paint:

TintMode

How tint color is applied to an image.

Traits

ImageResolver

Trait for resolving opaque image IDs to pixmaps at rasterization time.

Type Aliases

Color

A convenient alias for the color type used for Brush.

Image

An image.

PaintType

A kind of paint that can be used for filling and stroking shapes.