vello_common::paint

Type Alias Color

pub type Color = AlphaColor<Srgb>;

A convenient alias for the color type used for Brush.

Aliased Type

#[repr(transparent)]
pub struct Color {
    pub components: [f32; 4],
    pub cs: PhantomData<Srgb>,
}

Fields
components: [f32; 4]

The components, which may be manipulated directly.

The interpretation of the first three components depends on the color space. The fourth component is separate alpha.

cs: PhantomData<Srgb>

The color space.