vello_common::record

Trait Drawable

pub trait Drawable {
    // Required method
    fn bbox(&self, strips: &[Strip]) -> Option<RectU16>;
}

A drawable object that can report its bounding box.

Required Methods

fn bbox(&self, strips: &[Strip]) -> Option<RectU16>

Return the tile-aligned bounding box of the given object, if it has one.