vello_common::recording

Struct Recorder

pub struct Recorder<'a> {
    recording: &'a mut Recording,
    glyph_caches: Option<GlyphCaches>,
}

Recorder context that captures commands.

Fields
recording: &'a mut Recording

The recording to capture commands into.

glyph_caches: Option<GlyphCaches>
Implementations

impl<'a> Recorder<'a>

pub fn new( recording: &'a mut Recording, transform: Affine, glyph_caches: GlyphCaches, ) -> Self

Create a new recorder for the given recording.

pub fn fill_path(&mut self, path: &BezPath)

Fill a path with current paint and fill rule.

pub fn stroke_path(&mut self, path: &BezPath)

Stroke a path with current paint and stroke settings.

pub fn fill_rect(&mut self, rect: &Rect)

Fill a rectangle with current paint and fill rule.

pub fn stroke_rect(&mut self, rect: &Rect)

Stroke a rectangle with current paint and stroke settings.

pub fn set_transform(&mut self, transform: Affine)

Set the transform for subsequent operations.

pub fn set_fill_rule(&mut self, fill_rule: Fill)

Set the fill rule for subsequent fill operations.

pub fn set_stroke(&mut self, stroke: Stroke)

Set the stroke settings for subsequent stroke operations.

pub fn set_paint(&mut self, paint: impl Into<PaintType>)

Set the paint for subsequent rendering operations.

pub fn set_paint_transform(&mut self, paint_transform: Affine)

Set the current paint transform.

pub fn reset_paint_transform(&mut self)

Reset the current paint transform.

pub fn push_layer( &mut self, clip_path: Option<&BezPath>, blend_mode: Option<BlendMode>, opacity: Option<f32>, mask: Option<Mask>, )

Push a new layer with the given properties.

pub fn push_clip_layer(&mut self, clip_path: &BezPath)

Push a new clip layer.

pub fn pop_layer(&mut self)

Pop the last pushed layer.

pub fn glyph_run(&mut self, font: &FontData) -> GlyphRunBuilder<'_, Self>

Creates a builder for drawing a run of glyphs that have the same attributes.