vello_common::simd

Trait Splat4thExt

pub trait Splat4thExt<S> {
    // Required method
    fn splat_4th(self) -> Self;
}

Splatting every 4th element in the vector, used for splatting the alpha value of a color to all lanes.

Required Methods

fn splat_4th(self) -> Self

Splat every 4th element of the vector.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.