vello_common::tile

Struct Tile

#[repr(C)]
pub struct Tile {
    pub packed_winding_line_idx: u32,
    pub x: u16,
    pub y: u16,
}

A tile represents an aligned area on the pixmap, used to subdivide the viewport into sub-areas (currently 4x4) and analyze line intersections inside each such area.

Keep in mind that it is possible to have multiple tiles with the same index, namely if we have multiple lines crossing the same 4x4 area!

Note

This struct is #[repr(C)], but the byte order of its fields is dependent on the endianness of the compilation target.

Fields
packed_winding_line_idx: u32

The index of the line this tile belongs to into the line buffer, plus whether the line crosses the top edge of the tile, packed together.

The index is the unsigned number in the 31 least significant bits of this value.

The last bit is 1 if and only if the lines crosses the tileโs top edge. Lines making this crossing increment or decrement the coarse tile winding, depending on the line direction.

x: u16

The index of the tile in the x direction.

y: u16

The index of the tile in the y direction.

Implementations

impl Tile

pub const WIDTH: u16 = 4u16

The width of a tile in pixels.

pub const HEIGHT: u16 = 4u16

The height of a tile in pixels.

pub fn new_clamped(x: u16, y: u16, line_idx: u32, winding: bool) -> Self

Create a new tile. x and y will be clamped to the largest possible coordinate if they are too large.

line_idx must be smaller than MAX_LINES_PER_PATH.

pub(crate) const fn new(x: u16, y: u16, line_idx: u32, winding: bool) -> Self

pub const fn same_loc(&self, other: &Self) -> bool

Check whether two tiles are at the same location.

pub const fn prev_loc(&self, other: &Self) -> bool

Check whether self is adjacent to the left of other.

pub const fn same_row(&self, other: &Self) -> bool

Check whether two tiles are on the same row.

pub const fn line_idx(&self) -> u32

The index of the line this tile belongs to into the line buffer.

pub const fn winding(&self) -> bool

Whether the line crosses the top edge of the tile.

Lines making this crossing increment or decrement the coarse tile winding, depending on the line direction.

const fn to_bits(self) -> u64

Return the u64 representation of this tile.

This is the u64 interpretation of (y, x, packed_winding_line_idx) where y is the most-significant part of the number and packed_winding_line_idx the least significant.