vello_common

Module util

Utility functions.

Structs

Pool

Pool for reusing allocations.

RetainVec

A resizable vector that retains inner elements upon resizing.

Traits

Clear

A type that can be cleared.

Div255Ext

A trait for implementing a fast approximal division by 255 for integers.

RectExt

Extension methods for rectangles.

Functions

extract_scales

Extract scale factors from an affine transform using singular value decomposition.

f32_to_u8

Convert f32x16 to u8x16.

is_axis_aligned

Check if an affine transform has no skewing (i.e. preserves axis alignment).

is_integer_translation

Check if an affine transform is a pure integer translation.

normalized_mul_u8x16

Perform a normalized multiplication for u8x16.

normalized_mul_u8x32

Perform a normalized multiplication for u8x32.

snap_down 🔒

strip_bbox

Calculate the bounding box of the strips.

Type Aliases

VecPool

Pool for reusing vector allocations.