vello_common::encode

Trait EncodeExt

pub trait EncodeExt: Sealed {
    // Required method
    fn encode_into(
        &self,
        paints: &mut Vec<EncodedPaint>,
        transform: Affine,
    ) -> Paint;
}

A trait for encoding gradients.

Required Methods

fn encode_into( &self, paints: &mut Vec<EncodedPaint>, transform: Affine, ) -> Paint

Encode the gradient and push it into a vector of encoded paints, returning the corresponding paint in the process. This will also validate the gradient.