vello_common

Module paint

Types for paints.

Structs

ImageId

Opaque image handle

IndexedPaint

A paint that needs to be resolved via its index.

PremulColor

A premultiplied color.

Enums

ImageSource

Bitmap source used by Image.

Paint

A paint that is used internally by a rendering frontend to store how a wide tile command should be painted. There are only two types of paint:

Type Aliases

Image

An image.

PaintType

A kind of paint that can be used for filling and stroking shapes.